+++ keyboards/kprepublic/bm65hsrgb_iso/rev1/info.json
