+++ keyboards/kprepublic/bm65hsrgb_iso/rev1/keymaps/via/rules.mk
