+++ keyboards/kprepublic/bm65hsrgb_iso/rev1/rev1.c
@@ -13,7 +13,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "bm65hsrgb_iso.h"
+#include "rev1.h"
 
 #ifdef RGB_MATRIX_ENABLE
 led_config_t g_led_config = { {
