+++ keyboards/kprepublic/bm68hsrgb/rev1/info.json
