+++ keyboards/kprepublic/bm68hsrgb/rev1/keymaps/via/rules.mk
