+++ keyboards/kprepublic/bm68hsrgb/rev1/rev1.c
@@ -14,7 +14,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "bm68hsrgb.h"
+#include "rev1.h"
 
 #ifdef RGB_MATRIX_ENABLE
 led_config_t g_led_config = { {
